Move tuple_cat up into agency::
